HHH-9637 : Correction to test case for 4.2 branch
