+++ hibernate-entitymanager/src/test/java/org/hibernate/ejb/metamodel/AbstractMetamodelSpecificTest.java
@@ -32,6 +32,7 @@ public abstract class AbstractMetamodelSpecificTest extends BaseEntityManagerFun
 	public Class[] getAnnotatedClasses() {
 		return new Class[] {
 				Address.class, Alias.class, Country.class, CreditCard.class, Customer.class,
+				Entity1.class, Entity2.class, Entity3.class,
 				Info.class, LineItem.class, Order.class, Phone.class, Product.class,
 				ShelfLife.class, Spouse.class, Thing.class, ThingWithQuantity.class,
 				VersionedEntity.class
