+++ hibernate-entitymanager/src/test/java/org/hibernate/ejb/metamodel/Entity3.java
@@ -21,7 +21,7 @@
  * 51 Franklin Street, Fifth Floor
  * Boston, MA  02110-1301  USA
  */
-package org.hibernate.jpa.test.metamodel;
+package org.hibernate.ejb.metamodel;
 
 import javax.persistence.Entity;
 import javax.persistence.Id;
